fix: be explicit about which id in sql (#39)
Before Postgres was returning the movie id as id. Consequently, we were sometimes looking for reviews with the wrong id. Now, we're being explicit in the SQL. The only movie attribute we return with this SQL is movie_title, much like the other SQL queries in the reviewsRepo.

Also, I updated the test to ensure this bug 🐛 doesn't rear its ugly head again.